CRUD eventos

+++ app/Http/Controllers/AdminEventoController.php
@@ -9,6 +9,14 @@ use  App\Models\Ubicacion;
 class AdminEventoController extends Controller
 {
     public function crear(Request $request) {
+        $request->validate([
+            'titulo'  => 'required'     ,
+            'descripcion'  =>'required' ,
+            'fechaI'  =>  'required'  ,
+            'fechaF'  =>  'required' ,
+            'urlImagen'  =>   'required' ,
+            'tipo'  =>  'required'  ,
+        ]);
         $imagen = $request->file('urlImagen')->store('public/Evento');
         $urlFondoPath = Storage::url($imagen);
         $consulta= new Request([
@@ -38,40 +46,48 @@ class AdminEventoController extends Controller
             'titulo'  => $request['titulo']     ,
             'descripcion'  => $request['descripcion'],
             'fechaIni'  => $request['fechaI']   ,
-            'fechaFin'  => $request['fechaFin']   ,
-            'urlImg'  => ''   ,
-            'urlSecion'  => $request['enlace']  ,
+            'fechaFin'  => $request['fechaF']   ,
+            'urlImg'  => $evento['urlImg'] ,
+            'urlSecion'  => $request['enlaceE']  ,
             'tipo'  => $request['tipo']        ,
         ]);
         if($request->hasFile('urlImagen')){
             $imagen = $request->file('urlImagen')->store('public/Evento');
                 $urlFondoPath = Storage::url($imagen);
-                $consulta['urlImg'] = $urlFondoPath;
-            if($request['fechaIni'] == $evento['fechaIni'] && $request['fechaIni'] != null ){
-                $consulta['fechaIni'] = $evento['fechaIni'];
-            }
-            if($request['fechaFin'] == $evento['fechaFin'] && $request['fechaFin'] != null ){
-                $consulta['fechaFin'] = $evento['fechaFin'];
-            }
-            if($request['fechaFin'] == $evento['fechaFin'] && $request['fechaFin'] != null ){
-                $consulta['fechaFin'] = $evento['fechaFin'];
-            }
-            $evento = EventoController::create($consulta);
-            if($request['direccion'] != null){
-                $consulta = new Request([
-                    'direccion' => $request['direccion'],
-                    'tipo' => 5,
-                    'idEve' => $evento['id']
-                ]);
-                UbicacionController::create($consulta);
-            }
-            return redirect()->route('admin.indexEvento');
+                if($urlFondoPath != $evento['urlImg']) {
+                    $consulta['urlImg'] = $urlFondoPath;
+                }
         }
+        if($request['fechaI'] == null ){
+            $consulta['fechaIni'] = $evento['fechaIni'];
+        }
+        if($request['fechaF'] == null ){
+            $consulta['fechaFin'] = $evento['fechaFin'];
+        }
+        if($request['tipo'] == -1 ){
+            $consulta['tipo'] = $evento['tipo'];
+        }
+        if($request['enlaceE'] == null ){
+            $consulta['urlSecion'] = $evento['urlSecion'];
+        }
+        $evento = EventoController::update($consulta,$evento);
+        if($request['direccionE'] != null){
+            $ubicacion = Ubicacion::where('idEvento',$request['id'])->first();
+
+            if($ubicacion['direccion'] != $request['direccionE']){
+                $ubicacion->update([
+                    'direccion' => $request['direccionE'],
+                ]);
+            }
+        }
+        return redirect()->route('admin.indexEvento');
     }
 
     public function delete(Evento $evento) {
-        $ubicacion = Ubicacion::where('id',$evento->ubicaciones[0]->id)->first();
-        UbicacionController::delete($ubicacion);
+        if($evento->tipo == 2){
+            $ubicacion = Ubicacion::where('id',$evento->ubicaciones[0]->id)->first();
+            UbicacionController::delete($ubicacion);
+        }
         EventoController::delete($evento);
         return redirect()->route('admin.indexEvento');
     }
